refactor(overview): remove security findings banner (temporary) (refs #950)
Remove the Overview `🔒 Recent security findings` aggregate banner. The
ServiceDetails per-service security card remains the contextual surface and
the worker `securityAlerts` projection is untouched (data still flows).

Re-surfacing (Overview vs a dedicated screen) is deferred pending the #949
NVD first-party CVE data — issue stays open with the `tracking` label.

- Overview.jsx: drop the banner block + its `tagServiceForAlert` import and
  the now-unused `securityAlerts` destructure from usePolling
- en.js/ko.js: drop the unused `overview.security.title` key
- securityAlerts.js: mark `tagServiceForAlert` retained-but-unused for the
  deferred re-surfacing (its unit test stays)
- rename security-banner.spec.js → security-alerts.spec.js, rewritten to
  assert the retained ServiceDetails card (+ xAI non-leak assertion)
- CLAUDE.md: sync the security data-path note (no Overview banner)
